Borrar codigo duro de lab en user

+++ database/migrations/2020_03_21_011218_create_suspect_cases_table.php
@@ -64,7 +64,7 @@ class CreateSuspectCasesTable extends Migration
             $table->unsignedBigInteger('patient_id');
             $table->unsignedBigInteger('laboratory_id')->nullable();
 
-            $table->unsignedInteger('establishment_id');
+            $table->unsignedBigInteger('establishment_id');
 
             $table->timestamps();
             $table->softDeletes();
